Modularized my config. Added neo-tree, new tab plugin, and pyright.
