+++ lua/custom/plugins/indent.lua
@@ -1,0 +1,5 @@
+return {
+  -- 'tpope/vim-sleuth', -- Detect tabstop and shiftwidth automatically
+
+  'NMAC427/guess-indent.nvim',
+}
