+++ lua/custom/plugins/todo-comments.lua
@@ -1,0 +1,4 @@
+return {
+  -- Highlight todo, notes, etc in comments
+  { 'folke/todo-comments.nvim', event = 'VimEnter', dependencies = { 'nvim-lua/plenary.nvim' }, opts = { signs = false } },
+}
